requirements.txt
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
durationpy==0.9
Flask==3.1.0
Flask-Cors==5.0.0
google==3.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
kubernetes==32.0.0
MarkupSafe==3.0.2
node==1.2.2
npm==0.1.1
oauth2client==4.1.3
oauthlib==3.2.2
odict==1.9.0
optional-django==0.1.0
packaging==24.2
php==1.2.1
plumber==1.7
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylogrus==0.4.0
pymongo==4.11.1
pyparsing==3.2.1
python-bootstrap==0.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
shortuuid==1.0.13
six==1.17.0
soupsieve==2.6
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1